SoSo3.2
Streamlined menu, everything works properly (you plan to add highlighting of text, control arrows), not works correctly feature added, changed structure type to character strings, not fully correct implementation of input (reads last character). An error was detected in sorting, *fixed*, there is an error related to changing the structure type. The change function is also not fully debugged (the problem is the same as with adding). The input was changed to ASCII codes, which led to the addition of checks. Added comments to the structure and function "Output".