SoSo4
A competent conclusion was made, input was made, the structure was debugged, the types in the structure were deferred, sorting was postponed (the sorting selection menu does not work), the "change"function does not fully work. Deferred deletion. Enter the time separately, hours and minutes. The function with files remains (it is planned to start working with untyped files). There is also a problem with managing arrows. Graphical support via Ncurses.