SoSo<ncurses>v.3.4.1
I changed the menu for an adequate output and selection. The function call doesn't break now. There is a problem only in the sorting menu function (it does not show the sorting selection, by default it sorts in descending order). Added the Change function. The menu and graphics are output, but the values do not change yet (in search of an error).